+++ src/qmflows/hdf5/__init__.py
@@ -1,3 +1,0 @@
-from .quantumHDF5 import (StoreasHDF5, cp2k2hdf5)
-
-__all__ = ['StoreasHDF5', 'cp2k2hdf5']